More comprehensive support for attributes and their aliases in the community.general.ufw module. Broader use of omit to more accurately replicate the logic of community.general.ufw.
Modified logic for forming the direction attribute.
The default value 'in' is now assigned to non-routed
packets, allowing for a more concise definition of
forward packet rules in Ansible.
For example, the following now works:

 { rule: allow, port: 8123, route: True }

Previously, such a declaration would result in an Ansible error:
ERROR: Invalid interface clause for route rule.